컬럼 옵저버 CRUD 권한 제한

+++ src/colum/colum.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreateColumDto } from './dto/createColum.dto';
 import { UpdateColumDto } from './dto/updateColum.dto';
 import { In, Repository } from 'typeorm';
@@ -6,31 +6,39 @@ import { Colum } from './entities/colum.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import _, { isNil } from 'lodash';
 import { ReorderColumDto } from './dto/reorderColum.dto';
+import { User } from '../user/entity/user.entity';
+import { BoardMember } from '../board/entities/boardMember.entity';
 
 @Injectable()
 export class ColumService {
   constructor(
-    @InjectRepository(Colum) private readonly columRepository:Repository<Colum>
+    @InjectRepository(Colum) private readonly columRepository:Repository<Colum>,
+    @InjectRepository(BoardMember) private readonly boardMemberRepository: Repository<BoardMember> 
   ) {}
 
   //컬럼 생성
-  async create(createColumDto: CreateColumDto) {
+  async create(user: User, createColumDto: CreateColumDto) {
+    const {id: user_id} = user
     const {columOrder, board_id, title} = createColumDto
 
-  //   const existedColum = await this.columRepository
-  //       .createQueryBuilder('colum')
-  //       .leftJoinAndSelect("colum.board", "board.id")
-  //       .where("colum.title", {title})
-  //       .andWhere("board.id = boardId", {board_id})
-  //       .getOne()
+    const availableUser = await this.boardMemberRepository.findOne({
+      relations: ['user'],
+      where: {
+      user: {
+        id: user_id
+      }
+    }})
+    
+    if(availableUser.role == "observer") {
+      throw new UnauthorizedException('인가되지 않은 권합입니다.')
+    }
 
-  // if(existedColum) {
-  //   throw new BadRequestException('유효하지 않은 요청입니다.')
-  // }
 
     const colum = await this.columRepository.save({
       board_id, columOrder, title
     })
+
+
 
     return colum;
   }
@@ -38,39 +46,88 @@ export class ColumService {
 
 
   //컬럼 이름 수정
-  async update(updateColumDto: UpdateColumDto) {
+  async update(user: User, updateColumDto: UpdateColumDto) {
+    const {id: user_id} = user
     const {board_id, id, title} = updateColumDto
-    const existedColum = await this.columRepository.findOne({
-      relations: {
-        board: true
-      },
+    console.log(board_id, id, title)
+
+    const availableUser = await this.boardMemberRepository.findOne({
+      relations: ['user'],
       where: {
-        board: {
-          id: board_id,
-        },
-        title
+      user: {
+        id: user_id
+      }
+    }})
+
+    if(availableUser.role == "observer") {
+      throw new UnauthorizedException('인가되지 않은 권합입니다.')
+    }
+
+
+    const existedColum = await this.columRepository.findOne({
+      where: {
+        id,
+        board_id
+      }
+      })
+    if(_.isNil(existedColum)) {
+      throw new BadRequestException('[존재하지 않음]유효하지 않은 요청입니다.')
+    }
+
+    const existedColumName = await this.columRepository.findOne({
+      where: {
+        title,
+        board_id
       }
     })
-    if(existedColum) {
-      throw new BadRequestException('유효하지 않은 요청입니다.')
+    if(existedColumName) {
+      throw new BadRequestException('[중복된 제목]유효하지 않은 요청입니다.')
     }
-    
-    
+        
     return await this.columRepository.update({id}, {title})
   }
   //컬럼 삭제
-  async remove(id: number) {
+  async remove(user: User, id: number) {
+    const {id: user_id} = user
+
+    const availableUser = await this.boardMemberRepository.findOne({
+      relations: ['user'],
+      where: {
+      user: {
+        id: user_id
+      }
+    }})
+
+    if(availableUser.role == "observer") {
+      throw new UnauthorizedException('인가되지 않은 권합입니다.')
+    }
+
     const existedColum = await this.columRepository.findOneBy({id})
     if(isNil(existedColum)) {
       throw new BadRequestException('유효하지 않은 요청입니다.')
     }
+
 
     
     return await this.columRepository.delete({id})
   }
 
   //컬럼 순서 이동
-  async reorderColum({columIds, board_id}: ReorderColumDto) {
+  async reorderColum(user: User, {columIds, board_id}: ReorderColumDto) {
+    const {id: user_id} = user
+
+    const availableUser = await this.boardMemberRepository.findOne({
+      relations: ['user'],
+      where: {
+      user: {
+        id: user_id
+      }
+    }})
+
+    if(availableUser.role == "observer") {
+      throw new UnauthorizedException('인가되지 않은 권합입니다.')
+    }
+    
     const colums = await this.columRepository.find({
       where: {
         id: In(columIds)
@@ -79,6 +136,9 @@ export class ColumService {
     if(colums.length !== columIds.length) {
       throw new BadRequestException('유효하지 않은 요청입니다.')
     }
+
+
+
     for(let element of colums) {
       if(element.board_id !== board_id) {
         throw new BadRequestException('유효하지 않은 요청입니다.')
@@ -97,4 +157,6 @@ export class ColumService {
       })
     )
   }
+
+  ////////////////////////////////////////////////////////////
 }
